+++ src/Pages/main.js
@@ -342,7 +342,7 @@ function UserOptionMenu({ optionMenuState, setOptionMenuState }) {
     const status = await log_out();
 
     if (status === "OK") {
-      navigate("/Pages/login");
+      navigate("/");
     }
   }
