+++ src/Scripts/utils/user-log-out.js
@@ -1,8 +1,8 @@
 import Cookies from "js-cookie";
 
 export async function log_out() {
-    Cookies.remove('Username');
-    Cookies.remove('Token');
+    Cookies.remove('Username', { path: '' });
+    Cookies.remove('Token', { path: '' });
 
     return "OK";
 }
